fix(*): update the changes with the firebse integration pr
