+++ src/app/features/notification/notification.service.ts
@@ -18,6 +18,8 @@ import firebase from "firebase/compat/app";
 export class NotificationService {
   private messaging: any = null;
   readonly FIREBASE_CLOUD_MESSAGING_URL = `https://fcm.googleapis.com/v1/projects/${environment.firebaseConfig.projectId}/messages:send`;
+  private readonly NOTIFICATION_TOKEN_COOKIE_NAME = "notification_token";
+  private readonly DEVICE_NOTIFICATION_API_URL = "/api/v1/notification/device";
 
   constructor(private httpClient: HttpClient) {
     firebase.initializeApp(environment.firebaseConfig);
@@ -26,6 +28,7 @@ export class NotificationService {
 
   /**
    * Generates a consistent FCM payload.
+   * @param token - The FCM token for the device.
    * @param payload - Custom data payload.
    */
   private createFcmPayload(token: string, payload: Record<string, any> = {}) {
@@ -43,13 +46,26 @@ export class NotificationService {
    * Requests permission for Firebase messaging and retrieves the token.
    * Logs the token if successful, otherwise logs an error or info message.
    */
-  async getFcmToken(): Promise<any> {
+  async getFcmToken(): Promise<string> {
     try {
+      const existingToken = this.getNotificationTokenFromCookie();
+      if (existingToken) {
+        return existingToken;
+      }
+
       const notificationToken = await getToken(this.messaging, {
         vapidKey: environment.firebaseConfig.vapidKey,
       });
-      if (notificationToken) {
-        // TODO: Need to Implement the logic to save the FCM token in the Cookie and update in the backend API.
+
+      if (notificationToken !== existingToken) {
+        this.setCookie(
+          this.NOTIFICATION_TOKEN_COOKIE_NAME,
+          notificationToken,
+          30,
+        );
+
+        this.registerTokenWithBackend(notificationToken);
+
         return notificationToken;
       } else {
         Logging.log(
@@ -59,6 +75,45 @@ export class NotificationService {
     } catch (err) {
       Logging.error("An error occurred while retrieving token: ", err);
     }
+  }
+
+  /**
+   * Registers the device with the backend using the FCM token.
+   * @param notificationToken - The FCM token for the device.
+   */
+
+  private registerTokenWithBackend(notificationToken: string) {
+    const payload = { fcm_token: notificationToken };
+
+    try {
+      this.httpClient
+        .post(this.DEVICE_NOTIFICATION_API_URL, payload)
+        .subscribe();
+      Logging.log("Device registered successfully.");
+    } catch (err) {
+      Logging.error("Failed to register device: ", err);
+    }
+  }
+
+  /**
+   * Retrieves the value of a cookie by name.
+   * @param name - The name of the cookie.
+   */
+  getNotificationTokenFromCookie() {
+    const cookies = new URLSearchParams(document.cookie.replace(/; /g, "&"));
+    return cookies.get(this.NOTIFICATION_TOKEN_COOKIE_NAME) || null;
+  }
+
+  /**
+   * Sets a cookie with a specified name, value, and expiration days.
+   * @param name - The name of the cookie.
+   * @param value - The value of the cookie.
+   * @param days - The number of days until the cookie expires.
+   */
+  private setCookie(name: string, value: string, days: number): void {
+    const expires = new Date();
+    expires.setTime(expires.getTime() + days * 24 * 60 * 60 * 1000);
+    document.cookie = `${name}=${value};expires=${expires.toUTCString()};path=/`;
   }
 
   /**
@@ -83,7 +138,7 @@ export class NotificationService {
    * First retrieves the FCM token and then calls the notification sending function.
    */
   async sendNotification(payload: Record<string, any> = {}) {
-    const notificationToken = await this.getFcmToken();
+    const notificationToken = this.getNotificationTokenFromCookie();
     if (notificationToken) {
       const fcmPayload = this.createFcmPayload(notificationToken, payload);
       this.sendNotificationToUser(fcmPayload);
@@ -94,7 +149,7 @@ export class NotificationService {
 
   /**
    * Sends a notification using the FCM HTTP API.
-   * @param payload - The payload generated by createFcmPayload.
+   * @param fcmPayload - The payload generated by createFcmPayload.
    */
   sendNotificationToUser(fcmPayload: Object) {
     // TODO: Implement the FIREBASE_ACCESS_TOKEN from backend service
@@ -103,8 +158,10 @@ export class NotificationService {
       "Content-Type": "application/json",
     };
 
-    this.httpClient.post(this.FIREBASE_CLOUD_MESSAGING_URL, fcmPayload, {
-      headers,
-    });
+    this.httpClient
+      .post(this.FIREBASE_CLOUD_MESSAGING_URL, fcmPayload, {
+        headers,
+      })
+      .subscribe();
   }
 }
