--- src/app/notifications/notification.service.spec.ts
@@ -1,19 +1,0 @@
-import { TestBed } from "@angular/core/testing";
-import { provideHttpClientTesting } from "@angular/common/http/testing";
-import { provideHttpClient } from "@angular/common/http";
-import { NotificationService } from "./notification.service";
-
-describe("NotificationService", () => {
-  let service: NotificationService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [provideHttpClient(), provideHttpClientTesting()],
-    });
-    service = TestBed.inject(NotificationService);
-  });
-
-  it("should be created", () => {
-    expect(service).toBeTruthy();
-  });
-});
